// src/interfaces/display_properties.ts
export interface IDisplayProperties {
  BackgroundColor: number | undefined;
  Bold: boolean | undefined;
  FaceName: string | null;
  FontSize: number | undefined;
  HorizonalAlignment: number | undefined;
  Italic: boolean | undefined;
  Padding: number | undefined;
  TextColor: number | undefined;
  Underlined: boolean | undefined;
  VerticalALignment: number | undefined;
}